updated the script to match the current trunk
The sharing of Weld in JBoss AS currenly requires the JSF classes to be attached to the deployment rather than being shared. To that end, the original weld-core jar is split in two different parts while the JBoss AS is built. This fix replicates the process for the Weld build.