Issue Weld-2587 Jetty #3804
Implement a new style of integration with the jetty container that does not require access to jetty
private APIs (as enabled in the jetty cdi module).

Previous mechanism has been trimmed of deprecated support and retained as a legacy container that will
continue to work for users that do expose the jetty APIs (as enabled in the jetty cdi2 module).

Signed-off-by: Greg Wilkins <gregw@webtide.com>