Improve Invoker exception messages
If an invoker invocation causes a ClassCastException,
NullPointerException or IllegalArgumentexception, validate the instance
and arguments to see if the exception was caused by passing invalid
values. If so, discard the original exception and throw a new exception
with more information.

Doing these checks only when the invocation throws an exception avoids
the overhead of doing these checks before every successful invocation.